requirements.txt
cairosvg==2.8.2
decord==0.6.0
matplotlib==3.5.1
numpy==1.25.2
pandas==1.4.2
Pillow==9.0.1
Pillow==12.0.0
protobuf==6.33.2
qwen_vl_utils==0.0.14
scikit_learn==1.2.2
seaborn==0.11.2
sentence_transformers==5.2.0
torch==2.1.0
torchvision==0.14.1
tqdm==4.66.1
transformers==4.27.0.dev0
umap_learn==0.5.9.post2
